+++ CHANGELOG.md
@@ -7,6 +7,7 @@
     * add rsync client authentication with ssh key
     * add management of redis cluster with sentinel
     * add tunnel configuration for sql manager
+    * add method exec_statements in sql manager
 * Fixed ...
     * fixed parse_redis_uri to parse redis sentinel uri
     * replaced caracther | with + in simple.truncate